requirements.txt
streamlit==1.33.0
langchain==0.1.20
langchain-core==0.1.52
langchain-openai==0.1.7
openai==1.30.1
youtube-transcript-api==0.6.2
python-dotenv==1.0.0